adjust mock api response
