+++ __mocks__/api/user.ts
@@ -1,0 +1,29 @@
+import { rest } from "../rest";
+
+// Validate a token in the backend that is specific to a user
+const isValidatedAccessToken = (token?: string) => !!token;
+
+export const user = [
+  rest.get("/user/1", (req, res, ctx) => {
+    const token = req.headers.get("Authorization")?.split(" ").at(-1);
+
+    if (!isValidatedAccessToken(token)) {
+      return res(
+        ctx.status(401),
+        ctx.json({
+          error: "Invalid Access Token",
+        })
+      );
+    }
+
+    return res(
+      ctx.json({
+        id: 1,
+        email: "george.bluth@reqres.in",
+        first_name: "George",
+        last_name: "Bluth",
+        avatar: "https://reqres.in/img/faces/1-image.jpg",
+      })
+    );
+  }),
+];
