+++ src/main.tsx
@@ -4,8 +4,10 @@ import { App } from "@/App";
 
 const prepare = async (): Promise<void> => {
   if (import.meta.env.DEV && !import.meta.env.VITE_REACT_APP_API_HOST) {
-    const { mockServer: worker } = await import("@/__mocks__/browser");
-    worker.start();
+    const { mockServer } = await import("@/__mocks__/server");
+    mockServer.start({
+      onUnhandledRequest: "bypass",
+    });
   }
 };
